# api-contract.yaml
openapi: 3.0.3
info:
  title: CODELESS API
  version: 0.6.0
  description: >
    API contract for the CODELESS multi-model AI chat platform.
    Features: authentication (sign-up/login/logout), user profile with
    custom avatar and preferences, multi-turn chat with AI models
    (Server-Sent Events streaming & JSON fallback), model switching
    mid-conversation, admin management of PROVIDERS (CRUD, active toggle,
    per-provider default), admin management of MODELS (CRUD, active toggle,
    platform default, provider binding), admin global token quota limit
    per user, customizable system prompt, admin user listing & management
    with chat/token stats, admin dashboard KPI metrics, forwarding chat to
    real OpenAI-compatible upstreams, and standard OpenAI-compatible endpoints
    (/v1/models, /v1/chat/completions) behind bearer auth.
    All standard JSON responses are wrapped in a unified envelope:
    { success: boolean, message: string, data: any }.
  contact:
    name: CODELESS Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development API base
  - url: https://api.codeless.example.com/api/v1
    description: Production API base (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /admin/users:
    get:
      tags: [Admin - Users]
      summary: List all users with conversation counts and token usage
      description: Returns list of all registered users with their created conversations count and consumed tokens.
      responses:
        '200':
          description: List of users with statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersListApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Admin - Users]
      summary: Update user details, role, or reset/set used tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserAdminRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/status:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Admin - Users]
      summary: Enable or disable a user account without deleting it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: User status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin - Users]
      summary: Delete a user and cascade their conversations and messages
      description: Admins cannot delete their own account.
      responses:
        '204':
          description: User deleted successfully
        '400':
          description: Cannot delete self
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not allowed to perform this action (admin only)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
          description: Unique lowercase handle ([a-z0-9_]{3,30}).
        avatarUrl:
          type: string
          nullable: true
          description: Public URL served by GET /static/avatars/{userId}/{file}.
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
    UserProfileApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'
...
    AdminUserItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        isActive:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        usedTokens:
          type: integer
        conversationsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminUsersListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AdminUserItem'
    UpdateUserAdminRequest:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        displayName:
          type: string
        email:
          type: string
          format: email
        usedTokens:
          type: integer
          minimum: 0
    UpdateUserStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
...
    Error:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Invalid email or password"
        data:
          nullable: true
          example: null
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
    ApiResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          nullable: true
          description: Response payload
...